bug fix build x3
